requirements.txt
newsapi-python==0.2.7
spacy==3.7.4
neo4j==5.18.0
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38
sentence-transformers==2.7.0
transformers==4.40.1
torch==2.3.0
streamlit==1.33.0
pyngrok==7.1.6
python-dotenv==1.0.1
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl